Removing override of a method with a default implementation. (#227)
The word "Default" in the "DefaultSchemasCallback.java" let me assume
that I can derive classes from it to implement some method which are
needed in my use case. The fact that getAttribute is overridden in the
DefaultSchemasCallback.java leads to confusion and makes it even harder
to make it work with a derived class.
I think it is better to just let the default method in the interface
shining through.

Co-authored-by: Hristo Borisov <hristo.borisov@sap.com>
(cherry picked from commit 81fd5ffbaa85d2cfc704121a15dce299abfc3848)